requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
aiomysql==0.0.21
aioredis==1.3.1
amqp==5.0.6
async-timeout==3.0.1
attrs==21.2.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.8
esdk-obs-python==3.21.8
et-xmlfile==1.1.0
hiredis==2.0.0
httptools==0.3.0
idna==3.2
iniconfig==1.1.1
jsonschema==4.0.1
kafka-python==2.0.2
kombu==5.1.0
loguru==0.5.3
motor==2.5.1
multidict==5.1.0
Naked==0.1.31
openpyxl==3.0.9
packaging==21.0
peewee==3.14.4
peewee-async==0.7.1
pluggy==1.0.0
prompt-toolkit==3.0.20
py==1.10.0
pycparser==2.20
pycryptodome==3.10.1
pymongo==3.12.0
PyMySQL==0.9.3
pyparsing==3.0.1
pypinyin==0.44.0
pyrsistent==0.18.0
pytest==6.2.5
pytz==2021.3
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
sanic==21.6.2
sanic-routing==0.7.1
shellescape==3.8.1
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.2
ujson==4.2.0
urllib3==1.26.7
uvloop==0.16.0
vine==5.0.0
wcwidth==0.2.5
websockets==10.0
xmltodict==0.12.0
yarl==1.6.3
